+++ src/gdal_memfile.cpp
@@ -130,7 +130,9 @@ NAN_METHOD(Memfile::vsimemAnonymous) {
  * Delete and retrieve the contents of an in-memory /vsimem/ file.
  * This is a very fast zero-copy operation.
  * It does not block the event loop.
- * The file must not be open and there is no safety mechanism to enforce this.
+ * WARNING!
+ * The file must not be open or random memory corruption is possible with GDAL <= 3.3.1.
+ * GDAL >= 3.3.2 will fail gracefully further operations and this function is safe.
  * If the file was created by vsimem.set, it will return a reference
  * to the same Buffer that was used to create it.
  * Otherwise it will construct a new Buffer object with the GDAL