+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [3.4.0] WIP
+
+### Added
+ - PROJ 8.1.1
+ - GEOS 3.9.1
+ - GDAL 3.4.0-git
 
 ## [3.3.4] 2021-09-28